+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/media/InputPollOptionMedia.kt
@@ -27,6 +27,7 @@ object InputPollOptionMediaSerializer : KSerializer<InputPollOptionMedia> {
     override fun serialize(encoder: Encoder, value: InputPollOptionMedia) {
         when (value) {
             is TelegramMediaAnimation -> TelegramMediaAnimation.serializer().serialize(encoder, value)
+            is TelegramMediaLink -> TelegramMediaLink.serializer().serialize(encoder, value)
             is TelegramMediaLivePhoto -> TelegramMediaLivePhoto.serializer().serialize(encoder, value)
             is TelegramMediaLocation -> TelegramMediaLocation.serializer().serialize(encoder, value)
             is TelegramMediaPhoto -> TelegramMediaPhoto.serializer().serialize(encoder, value)